Bundle helper functions into util module
