+++ util/settings.py
@@ -1,9 +1,12 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
 
-from xbmc import translatePath
+import re
 
-import utils
+from xbmcvfs import translatePath
+
+from util.addon_info import ADDON
+from util.logging import kodi
 
 # Exhaustive list of constants as used by the addon's settings
 service_enabled = "service_enabled"
@@ -20,7 +23,7 @@ debugging_enabled = "debugging_enabled"
 
 default_action = "default_action"
 cleaning_type = "cleaning_type"
-clean_kodi_library = "clean_kodi_library"
+clean_library = "clean_library"
 clean_movies = "clean_movies"
 clean_tv_shows = "clean_tv_shows"
 clean_music_videos = "clean_music_videos"
@@ -52,13 +55,34 @@ exclusion4 = "exclusion4"
 exclusion5 = "exclusion5"
 
 bools = [service_enabled, delete_folders, clean_related, notifications_enabled, notify_when_idle, debugging_enabled,
-         clean_kodi_library, clean_movies, clean_tv_shows, clean_music_videos, clean_when_idle, enable_expiration,
+         clean_library, clean_movies, clean_tv_shows, clean_music_videos, clean_when_idle, enable_expiration,
          clean_when_low_rated, ignore_no_rating, clean_when_low_disk_space, create_subdirs,
          not_in_progress, keep_hard_linked, exclusion_enabled]
 strings = [ignore_extensions, cleaning_type, default_action]
 numbers = [delayed_start, scan_interval, expire_after, minimum_rating, disk_space_threshold]
 paths = [disk_space_check_path, holding_folder, create_subdirs, exclusion1, exclusion2, exclusion3, exclusion4,
          exclusion5]
+
+
+def anonymize_path(path):
+    """
+    :type path: unicode
+    :param path: The network path containing credentials that need to be stripped.
+    :rtype: unicode
+    :return: The network path without the credentials.
+    """
+    if "://" in path:
+        kodi.debug(f"Anonymizing {path}")
+        # Look for anything matching a protocol followed by credentials
+        # This regex assumes there is no @ in the remainder of the path
+        regex = "^(?P<protocol>smb|nfs|afp|upnp|http|https):\/\/(.+:.+@)?(?P<path>[^@]+?)$"
+        results = re.match(regex, path, flags=re.I | re.U).groupdict()
+
+        # Keep only the protocol and the actual path
+        path = f"{results['protocol']}://{results['path']}"
+        kodi.debug(f"Result: {path}")
+
+    return path
 
 
 def get_value(setting):
@@ -71,13 +95,13 @@ def get_value(setting):
     :return: The value corresponding to the provided setting. This can be a float, a bool, a string or None.
     """
     if setting in bools:
-        return bool(utils.ADDON.getSetting(setting) == "true")
+        return bool(ADDON.getSetting(setting) == "true")
     elif setting in numbers:
-        return float(utils.ADDON.getSetting(setting))
+        return float(ADDON.getSetting(setting))
     elif setting in strings:
-        return str(utils.ADDON.getSetting(setting))
+        return str(ADDON.getSetting(setting))
     elif setting in paths:
-        return utils.anonymize_path(str(translatePath(utils.ADDON.getSetting(setting))))
+        return anonymize_path(str(translatePath(ADDON.getSetting(setting))))
     else:
         raise ValueError(f"Failed loading {setting} value. Type {type(setting)} cannot be handled.")
